Add site.pp from puppetlabs/openstack investigation.
This is based on the example site.pp in the openstack module, with a large
number of edits.  It contains one controller and one compute node.  If you
want to use this yourself, you will have to edit several IP addresses and
hostnames throughout the file: lines 65, 72, and 121.

Currently, VNC does not work.  There are a few more caveats, described in the
Puppet for Centos wiki page.

The networking setup is bass-ackwards right now, with an insane floating IP
range, for instance.  This will be fixed later.