Misc. cleanup
No significant content changes, just tweaks like:

* spelling/grammar fixes
* changed some things that scan poorly
* fixed some terminology (e.g. github's "pull" mechanism -> "pull
  request")

As well as some global stylistic changes, specifically:

* Hard line wrapping at 72 characters.
* Use of footnote-style links for long urls.

The primary design goal for markdown was to be readable in source form,
not just when rendered as HTML. These changes are to facilitate that.
Hard line wrapping helps when viewing a file via `cat` or similar, and
the footnotes prevent long urls from interrupting the reader's flow.